pyproject.toml
[project]
name = "timeseries-etl"
version = "0.1.0"
description = "Timeseries ETL pipeline for sensor data preparation, plotting, and report generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "duckdb>=0.9",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "scikit-learn>=1.3",
    "python-docx>=1.0",
    "docxcompose>=1.3",
    "python-dateutil>=2.8",
]

[project.scripts]
timeseries-etl = "timeseries_etl.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/timeseries_etl"]
